review_tag 페이지 위한 activity 파일 삭제
//review_tag_fragment로 이동, 아직 연결 불가
